# Reduce re-renders from block nodes context