@ellatrix 2020-05-06
@youknowriad Oh wow this is actually MUCH better. Look at the block selection time.

```
Ellas-MacBook-Pro:gutenberg ella$ git checkout master
Switched to branch 'master'
Your branch is up to date with 'origin/master'.
Ellas-MacBook-Pro:gutenberg ella$ npm run test-performance

> gutenberg@8.0.0 test-performance /Users/ella/gutenberg
> wp-scripts test-e2e --config packages/e2e-tests/jest.performance.config.js

Chromium is already in /Users/ella/gutenberg/node_modules/puppeteer/.local-chromium/mac-737027; skipping download.
 PASS  packages/e2e-tests/specs/performance/performance.test.js (47.127s)
  Performance
    ✓ Loading, typing and selecting blocks (41008ms)

Test Suites: 1 passed, 1 total
Tests:       1 passed, 1 total
Snapshots:   0 total
Time:        47.275s
Ran all test suites.

Loading Time:
Average time to load: 6101ms
Average time to DOM content load: 6070ms

Typing Performance:
Average time to type character: 25.54ms
Slowest time to type character: 64.87ms
Fastest time to type character: 19.69ms

Block Selection Performance:
Average time to select a block: 411.09ms
Slowest time to select a block: 942.86ms
Fastest time to select a block: 318.47ms

Ellas-MacBook-Pro:gutenberg ella$ git checkout -
Switched to branch 'fix/block-nodes-context'
Ellas-MacBook-Pro:gutenberg ella$ npm run test-performance

> gutenberg@8.0.0 test-performance /Users/ella/gutenberg
> wp-scripts test-e2e --config packages/e2e-tests/jest.performance.config.js

Chromium is already in /Users/ella/gutenberg/node_modules/puppeteer/.local-chromium/mac-737027; skipping download.
 PASS  packages/e2e-tests/specs/performance/performance.test.js (37.968s)
  Performance
    ✓ Loading, typing and selecting blocks (32825ms)

Test Suites: 1 passed, 1 total
Tests:       1 passed, 1 total
Snapshots:   0 total
Time:        38.113s, estimated 48s
Ran all test suites.

Loading Time:
Average time to load: 6086ms
Average time to DOM content load: 6057ms

Typing Performance:
Average time to type character: 24.12ms
Slowest time to type character: 66.53ms
Fastest time to type character: 20.04ms

Block Selection Performance:
Average time to select a block: 38.13ms
Slowest time to select a block: 58.94ms
Fastest time to select a block: 31.82ms
```