@ellatrix 2020-05-06
While there's a performance improvement on block selection, it doesn't change anything for #22099.