@youknowriad 2020-05-06
Are you sure about the block selection performance impact here? because  it's huuge