@ellatrix 2020-05-06
New test with a rebase to master. Takes now only 10% of the time... I expected some improvement but not at all _that_ much.

master:

```
Loading Time:
Average time to load: 5953ms
Average time to DOM content load: 5924ms

Typing Performance:
Average time to type character: 23.83ms
Slowest time to type character: 59.35ms
Fastest time to type character: 18.07ms

Block Selection Performance:
Average time to select a block: 345.45ms
Slowest time to select a block: 450.2ms
Fastest time to select a block: 311.45ms
```

pr:

```
Loading Time:
Average time to load: 5805ms
Average time to DOM content load: 5777ms

Typing Performance:
Average time to type character: 24.08ms
Slowest time to type character: 69.43ms
Fastest time to type character: 18.66ms

Block Selection Performance:
Average time to select a block: 40.35ms
Slowest time to select a block: 55.16ms
Fastest time to select a block: 34.94ms
```