@ellatrix 2020-05-06
This also seems to fix an intermittent e2e test failure in `master`, unless it's passed here a few by accident. Seems to fail a lot on `master` though.