@ellatrix 2020-05-06
Sounds good. Let's see how we're going to use it in other places.Also if it could be useful for plugins.